Interfaces, implementation classes and test classes added.
