--- src/main/java/uk/gov/hmcts/reform/roleassignment/data/resource/RoleAssignmentHistoryRepository.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.data.resource;
-
-public interface RoleAssignmentHistoryRepository {
-}
