--- src/main/java/uk/gov/hmcts/reform/roleassignment/data/resource/RoleAssignmentRepository.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.data.resource;
-
-public interface RoleAssignmentRepository {
-}
