--- src/main/java/uk/gov/hmcts/reform/roleassignment/data/resourcerequest/RoleAssignmentRequestRepository.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.data.resourcerequest;
-
-public interface RoleAssignmentRequestRepository {
-}
