--- src/main/java/uk/gov/hmcts/reform/roleassignment/data/resourcerequest/RoleAssignmentRequestStatusRepository.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.data.resourcerequest;
-
-public interface RoleAssignmentRequestStatusRepository {
-}
