+++ src/main/java/uk/gov/hmcts/reform/roleassignment/data/roleassignment/RoleAssignmentHistoryRepository.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.data.roleassignment;
+
+public interface RoleAssignmentHistoryRepository {
+}
