+++ src/main/java/uk/gov/hmcts/reform/roleassignment/data/roleassignment/RoleAssignmentRepository.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.data.roleassignment;
+
+public interface RoleAssignmentRepository {
+}
