+++ src/main/java/uk/gov/hmcts/reform/roleassignment/data/rolerequest/RoleAssignmentRequestRepository.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.data.rolerequest;
+
+public interface RoleAssignmentRequestRepository {
+}
