+++ src/main/java/uk/gov/hmcts/reform/roleassignment/data/rolerequest/RoleAssignmentRequestStatusRepository.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.data.rolerequest;
+
+public interface RoleAssignmentRequestStatusRepository {
+}
